requirements.txt
django
pandas
liac-arff
gunicorn
scikit-learn
